refactor(bin): make `calculate_hash` a utility function
