+++ powersoftau/src/bin/compute_constrained.rs
@@ -1,14 +1,15 @@
-use powersoftau::batched_accumulator::BatchedAccumulator;
-use powersoftau::keypair::keypair;
-use powersoftau::parameters::{CheckForCorrectness, UseCompression};
+use powersoftau::{
+    batched_accumulator::BatchedAccumulator,
+    keypair::keypair,
+    parameters::{CeremonyParams, CheckForCorrectness, CurveKind, UseCompression},
+    utils::calculate_hash,
+};
 
 use bellman_ce::pairing::bn256::Bn256;
 use memmap::*;
 use std::fs::OpenOptions;
 
 use std::io::{Read, Write};
-
-use powersoftau::parameters::{CeremonyParams, CurveKind};
 
 const INPUT_IS_COMPRESSED: UseCompression = UseCompression::No;
 const COMPRESS_THE_OUTPUT: UseCompression = UseCompression::Yes;
@@ -136,7 +137,7 @@ fn main() {
         UseCompression::No == INPUT_IS_COMPRESSED,
         "Hashing the compressed file in not yet defined"
     );
-    let current_accumulator_hash = BatchedAccumulator::<Bn256>::calculate_hash(&readable_map);
+    let current_accumulator_hash = calculate_hash(&readable_map);
 
     {
         println!("`challenge` file contains decompressed points and has a hash:");
@@ -213,7 +214,7 @@ fn main() {
     let output_readonly = writable_map
         .make_read_only()
         .expect("must make a map readonly");
-    let contribution_hash = BatchedAccumulator::<Bn256>::calculate_hash(&output_readonly);
+    let contribution_hash = calculate_hash(&output_readonly);
 
     print!(
         "Done!\n\n\
