+++ powersoftau/src/bin/reduce_powers.rs
@@ -2,7 +2,7 @@ use bellman_ce::pairing::bn256::Bn256;
 use powersoftau::{
     batched_accumulator::BatchedAccumulator,
     parameters::{CeremonyParams, CheckForCorrectness, CurveKind, UseCompression},
-    utils::reduced_hash,
+    utils::{calculate_hash, reduced_hash},
 };
 
 use std::fs::OpenOptions;
@@ -105,7 +105,7 @@ fn main() {
     let output_readonly = writable_map
         .make_read_only()
         .expect("must make a map readonly");
-    let contribution_hash = BatchedAccumulator::<Bn256>::calculate_hash(&output_readonly);
+    let contribution_hash = calculate_hash(&output_readonly);
 
     println!("Reduced contribution is formed with a hash:");
